Backend token

+++ src/services/application.service.ts
@@ -2,51 +2,77 @@ import * as R from "ramda";
 import { type Application } from "../models/application.model";
 import { APPLICATION_REPOSITORY, httpError } from "../utils/const";
 import { ServiceError } from "../utils/service-errors";
-import type Context from "../middlewares/context.ts";
+import type Context from "../utils/context.ts";
 import type { IApplicationRepository } from "../repositories/interfaces.ts";
 import { type Environment } from "../models/environment.model.ts";
+import { type Token } from "../models/token.model.ts";
 import { PrismaClientKnownRequestError } from "@prisma/client/runtime/library";
+import type { BackendTokenPermissions } from "../utils/types.ts";
 
 const getApplication = async ({
   context,
   appName,
   clerkId,
+  tokenPermissions,
 }: {
   context: Context;
   appName: string;
-  clerkId: string;
+  clerkId?: string;
+  tokenPermissions?: BackendTokenPermissions;
 }): Promise<
   Omit<Application, "environments"> & {
     environments: Pick<Environment, "id" | "name" | "description">[];
   }
 > => {
+  if (!clerkId && !tokenPermissions) {
+    throw new ServiceError(httpError.USER_NOT_AUTHENTICATED);
+  }
+
   const repository = context.get<IApplicationRepository>(
     APPLICATION_REPOSITORY,
   );
 
-  const application = await repository.getApplication({ appName, clerkId });
+  const application = await repository.getApplication({
+    appName,
+    clerkId,
+    tokenPermissions,
+  });
   if (!application) {
     throw new ServiceError(httpError.APPNAME_NOT_FOUND);
   }
   return application;
 };
 
-const getUserApplications = async ({
+const getApplications = async ({
   context,
   clerkId,
+  tokenPermissions,
 }: {
   context: Context;
-  clerkId: string;
+  clerkId?: string;
+  tokenPermissions?: BackendTokenPermissions;
 }): Promise<
   (Omit<Application, "id" | "environments"> & {
     environments: Omit<Environment, "id" | "description">[];
   })[]
 > => {
+  if (!clerkId && !tokenPermissions) {
+    throw new ServiceError(httpError.USER_NOT_AUTHENTICATED);
+  }
+
   const repository = context.get<IApplicationRepository>(
     APPLICATION_REPOSITORY,
   );
 
-  return repository.getUserApplications({ clerkId });
+  if (clerkId) {
+    return repository.getUserApplications({ clerkId });
+  } else {
+    const app = await repository.getTokenApplication({
+      tokenPermissions: tokenPermissions!,
+    });
+    if (!app) return [];
+    return [app];
+  }
 };
 
 const createApplication = async ({
@@ -55,24 +81,36 @@ const createApplication = async ({
   clerkId,
   image,
   appDescription,
+  environmentName,
+  environmentDescription,
 }: {
   context: Context;
   appName: string;
-  clerkId: string;
+  clerkId?: string;
   image: string;
   appDescription: string;
-}): Promise<void> => {
+  environmentName?: string;
+  environmentDescription?: string;
+}): Promise<{
+  applicationName: string;
+  environmentName: string;
+  applicationTokens: Token[];
+  environmentTokens: Token[];
+}> => {
   const repository = context.get<IApplicationRepository>(
     APPLICATION_REPOSITORY,
   );
 
   try {
-    await repository.createApplication({
+    const result = await repository.createApplication({
       appName,
       clerkId,
       image,
       appDescription,
+      environmentName,
+      environmentDescription,
     });
+    return result;
   } catch (e: any) {
     if (e instanceof PrismaClientKnownRequestError && e.code === "P2002") {
       throw new ServiceError(httpError.APPNAME_EXISTS);
@@ -89,10 +127,14 @@ const updateApplication = async (props: {
   context: Context;
   oldAppName: string;
   newAppName?: string;
-  clerkId: string;
+  clerkId?: string;
+  tokenPermissions?: BackendTokenPermissions;
   description?: string;
   image?: string;
-}): Promise<Omit<Application, "environments"> | null> => {
+}): Promise<Omit<Application, "environments" | "tokens"> | null> => {
+  if (!props.clerkId && !props.tokenPermissions) {
+    throw new ServiceError(httpError.USER_NOT_AUTHENTICATED);
+  }
   const repository = props.context.get<IApplicationRepository>(
     APPLICATION_REPOSITORY,
   );
@@ -107,6 +149,7 @@ const updateApplication = async (props: {
     return await repository.updateApplication({
       oldAppName: props.oldAppName,
       clerkId: props.clerkId,
+      token: props.tokenPermissions?.token,
       updateProps,
     });
   } catch (e) {
@@ -125,23 +168,31 @@ const deleteApplication = async ({
   context,
   clerkId,
   appName,
+  tokenPermissions,
 }: {
   context: Context;
   appName: string;
-  clerkId: string;
+  clerkId?: string;
+  tokenPermissions?: BackendTokenPermissions;
 }): Promise<Omit<Application, "environments"> | null> => {
+  if (!clerkId && !tokenPermissions) {
+    throw new ServiceError(httpError.USER_NOT_AUTHENTICATED);
+  }
+
   const repository = context.get<IApplicationRepository>(
     APPLICATION_REPOSITORY,
   );
-  const application = await repository.getApplication({ appName, clerkId });
+  const application = await repository.getApplication({
+    appName,
+    clerkId,
+    tokenPermissions,
+  });
   if (!application) {
     throw new ServiceError(httpError.APP_DOESNT_EXIST);
   }
 
   try {
     return repository.deleteApplication({
-      appName,
-      clerkId,
       dbAppId: application.id,
     });
   } catch (e) {
@@ -152,7 +203,7 @@ const deleteApplication = async ({
 
 export {
   getApplication,
-  getUserApplications,
+  getApplications,
   createApplication,
   updateApplication,
   deleteApplication,
